+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+- Add datasets for the real-robot stage
 
 ## [1.0.1] - 2022-07-04
 ### Changed